chore: Move Image types to seperate file
