--- stranger_danger/classifier/__init__.py
@@ -1,4 +1,0 @@
-from stranger_danger.classifier.cv2dnn.cv2_dnn import Cv2Dnn
-from stranger_danger.classifier.protocol import Classifier, Prediction, Predictions
-
-__all__ = ["Cv2Dnn", "Classifier", "Prediction", "Predictions"]
