+++ stranger_danger/constants/image_types.py
@@ -1,0 +1,6 @@
+from typing import NewType
+
+import numpy as np
+
+AnnotadedImage = NewType("AnnotadedImage", np.ndarray)
+FenceImage = NewType("FenceImage", np.ndarray)
