--- stranger_danger/fences/__init__.py
@@ -1,5 +1,0 @@
-from stranger_danger.fences.circular_fence import CircularFence
-from stranger_danger.fences.protocol import Coordinate, Fence
-from stranger_danger.fences.rectangular_fence import RectangularFence
-
-__all__ = ["CircularFence", "Coordinate", "Fence", "RectangularFence"]
